Rename files as requested by reviewers
