+++ desc/integrals/_interp_utils.py
@@ -14,7 +14,7 @@ from interpax import interp1d
 from orthax.chebyshev import chebroots
 
 from desc.backend import dct, jnp, rfft, rfft2, take
-from desc.integrals.quad_utils import bijection_from_disc
+from desc.integrals._quad_utils import bijection_from_disc
 from desc.utils import Index, errorif, safediv
 
 # TODO:
