+++ desc/integrals/_quad_utils.py
